refactor: rename kafka_controller.rs -> controller.rs
