+++ rust/operator-binary/src/discovery.rs
@@ -10,8 +10,8 @@ use stackable_operator::{
 };
 
 use crate::{
+    controller::KAFKA_CONTROLLER_NAME,
     crd::{role::KafkaRole, security::KafkaTlsSecurity, v1alpha1},
-    kafka_controller::KAFKA_CONTROLLER_NAME,
     utils::build_recommended_labels,
 };
