+++ rust/operator-binary/src/resource/listener.rs
@@ -5,8 +5,8 @@ use stackable_operator::{
 };
 
 use crate::{
+    controller::KAFKA_CONTROLLER_NAME,
     crd::{role::broker::BrokerConfig, security::KafkaTlsSecurity, v1alpha1},
-    kafka_controller::KAFKA_CONTROLLER_NAME,
     utils::build_recommended_labels,
 };
