+++ rust/operator-binary/src/resource/statefulset.rs
@@ -48,6 +48,7 @@ use crate::{
         command::{broker_kafka_container_commands, controller_kafka_container_command},
         node_id_hasher::node_id_hash32_offset,
     },
+    controller::KAFKA_CONTROLLER_NAME,
     crd::{
         self, APP_NAME, KAFKA_HEAP_OPTS, LISTENER_BOOTSTRAP_VOLUME_NAME,
         LISTENER_BROKER_VOLUME_NAME, LOG_DIRS_VOLUME_NAME, METRICS_PORT, METRICS_PORT_NAME,
@@ -60,7 +61,6 @@ use crate::{
         security::KafkaTlsSecurity,
         v1alpha1,
     },
-    kafka_controller::KAFKA_CONTROLLER_NAME,
     kerberos::add_kerberos_pod_config,
     operations::graceful_shutdown::add_graceful_shutdown_config,
     product_logging::{
